% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_tstat.R
\name{adj_tstat}
\alias{adj_tstat}
\title{adj_tstat}
\usage{
adj_tstat(dt, quant, block_size, gps_mat, covs)
}
\arguments{
\item{dt}{Data.frame object in R. This data frame must have quant and covs as variable names}

\item{quant}{Character value identifying the categorical values for the treatment variables}

\item{block_size}{Scalar value that defines the number of blocks to create based on the GPS values evaluated at the median of the quantile. The default is five blocks.}

\item{gps_mat}{Matrix with the generalized propensity score values for each individual evaluated at the median of each treatment interval}

\item{covs}{Character value or vector that contains the covariate to check for balance}
}
\value{
List of objects
}
\description{
This function allows the user to calculate the t-test statistics adjusted for the Generalized Propensity Score.
It follows the development of Hirano and Imbens (2004).
}
