% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_nav.R
\name{continuous_nav}
\alias{continuous_nav}
\title{Estimating the Dose Response Function given the Generalized Propensity Score}
\usage{
continuous_nav(formula, data, intervals = 50)
}
\arguments{
\item{formula}{This is a formula object specifying the dose-response model}

\item{data}{This is the data frame used to fit the model}

\item{intervals}{Specify the number of intervals to include.}
}
\description{
This function estimates the fitted value from a linear hierarchical mixed effects model, with the
generalized propensiy score and treatment occuring at the highest level
}
