% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_functions.R
\name{normal_gps}
\alias{normal_gps}
\title{Generalized Propensity Score for Continuous Treatment Domain}
\usage{
normal_gps(tx, covs, gps_val = NULL, interact_vars = NULL,
  polynomial_vars = NULL, polynomial_deg = NULL, variable_selection = F)
}
\arguments{
\item{tx}{Vector with the continuous treatment value, used for finding the initial parameter MLE's}

\item{covs}{Matrix of observed covariates}

\item{gps_val}{Scalar value or vector which contains the values to find the estimated generalized propensity score}

\item{interact_vars}{Specifies a character subset of the variables from the matrix covs and adds in pairwise interactions between all included covariates}

\item{polynomial_vars}{Specifies a character subset of variables from the matrix covs to include as polynomial terms}

\item{polynomial_deg}{Scalar value that identifies to what power the polynomial variable should be raised}

\item{variable_selection}{Indicator whether to perform variable selection using AIC backwards selection}
}
\value{
Returns a list of objects.
}
\description{
This function is used to fit a normal model for the conditional distribution of the treatment given covariates, and returns the resulting score function values for the observed treatment. After fitting the observed values the user can specify specific fixed treatment values to evaluate the conditional density at these points.
}
