% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_create.R
\name{quant_create}
\alias{quant_create}
\title{Quantile Creation Function}
\usage{
quant_create(continuous_var, n, dt = NULL)
}
\arguments{
\item{continuous_var}{If dt is specified, then this should be a character variable specifying which variable to split on. If dt is null then this should be a continuous numeric vector.}

\item{n}{The number of evenly spaced quantiles to create}

\item{dt}{Data.fame object in R, does not need to be specified if supplying a continuous vector}
}
\value{
List containing the vector of quantile values and the continuous cutpoints used.
}
\description{
A simple function that allows you to create evenly spaced quantiles for a continuous varible.
}
\examples{
data("iris")
quant_obj <- quant_create(iris$Sepal.Length)
table(quant_obj$quant); quant_obj$cut

}
