% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_functions.R
\name{R_score}
\alias{R_score}
\title{Generalized Propensity Score for Continuous Treatment Domain}
\usage{
R_score(tx, X, beta, sigma)
}
\description{
Generalized Propensity Score for Continuous Treatment Domain
}
