% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{t_test}
\usage{
t_test(dt, covariates, tx = NULL, tx_cat = NULL)
}
\arguments{
\item{dt}{Data.frame object in R. This data frame must have covariates and tx_var as variable names}

\item{tx_cat}{Character value identifying the categorical values for the treatment variables}
}
\value{
List of objects containing \code{mean_table}, \code{t_table}, and \code{obs}
}
\description{
This is a function that will test balance of the covariates with respect to a
continuous treatment variable. In a non-randomized experiment, such as an observational experiment,
we expect that certain covariates will be highly correlated with the treatment and potentially the outcome.
We can use this function to gauge the level of imbalance in the covariates.
}
\details{
The t-tests are run comparing one reference category of the continuous variable against the remaining categories.
By default the continuous variable will be
}
