% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_function.R
\name{wd_choice}
\alias{wd_choice}
\title{Working Directory Function}
\usage{
wd_choice(comp_choice, type_save)
}
\arguments{
\item{comp_choice}{Character value specifying which of the three computers I am using}

\item{type_save}{Character value specifying which type of file to save, e.g. Image}
}
\value{
Returns the working directory of interest
}
\description{
Working Directory Function
}
